+++ crates/client/rpc/src/methods/read/get_class.rs
@@ -30,9 +30,12 @@ pub fn get_class(_block_id: BlockId, class_hash: FieldElement) -> RpcResult<Cont
     };
 
     // converting from stored Blockifier class to rpc class
-    let StorageContractClassData { contract_class, abi } = contract_class_data;
-    Ok(ContractClassWrapper { contract: contract_class, abi }.try_into().map_err(|e| {
-        log::error!("Failed to convert contract class from hash '{class_hash}' to RPC contract class: {e}");
-        StarknetRpcApiError::InternalServerError
-    })?)
+    // TODO: retrieve sierra_program_length and abi_length when they are stored in the storage
+    let StorageContractClassData { contract_class, abi, sierra_program_length, abi_length } = contract_class_data;
+    Ok(ContractClassWrapper { contract: contract_class, abi, sierra_program_length, abi_length }.try_into().map_err(
+        |e| {
+            log::error!("Failed to convert contract class from hash '{class_hash}' to RPC contract class: {e}");
+            StarknetRpcApiError::InternalServerError
+        },
+    )?)
 }