+++ crates/primitives/contract/src/class.rs
@@ -16,8 +16,8 @@ use alloc::vec::Vec;
 pub struct StorageContractClassData {
     pub contract_class: ContractClassBlockifier,
     pub abi: ContractAbi,
-    // pub sierra_program_length: u64,
-    // pub abi_length: u64,
+    pub sierra_program_length: u64,
+    pub abi_length: u64,
 }
 
 #[derive(Debug, Clone)]
@@ -43,6 +43,8 @@ pub struct ContractClassData {
 pub struct ContractClassWrapper {
     pub contract: ContractClassBlockifier,
     pub abi: ContractAbi,
+    pub sierra_program_length: u64,
+    pub abi_length: u64,
 }
 
 #[cfg(feature = "std")]
@@ -75,7 +77,13 @@ pub mod convert {
                 }
             };
 
-            Ok(Self { contract, abi })
+            let sierra_program_length = match contract_class {
+                ContractClassCore::Sierra(class_sierra) => class_sierra.sierra_program.len(),
+                ContractClassCore::Legacy(_) => 0,
+            } as u64;
+            let abi_length = abi.length() as u64;
+
+            Ok(Self { contract, abi, sierra_program_length, abi_length })
         }
     }